requirements.txt
Flask~=2.3.2
Gunicorn
SQLAlchemy~=2.0.25
PyMySql
python-dotenv~=1.0.0